x86/paravirt: allow paravirt-ops patching post-init
paravirt_ops patching is used to accelerate ops by annotating them at
compile time (via PVOP.* macros) such that all their callsites can be
patched in the kernel (and modules) at init so as to avoid the cost of
indirect calls.  In some cases, the target op is inlined at the
callsite.

To allow this to happen post-init, we now persist the subset of
interest in .parainstructions.runtime.

Signed-off-by: Ankur Arora <ankur.a.arora@oracle.com>